feat: Recolte validation

+++ Citronix/Citronix/src/main/java/com/app/Citronix/Repository/ChampRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.app.Citronix.Model.Entity.Champ;
 
 @Repository
-public interface ChampRepository extends JpaRepository<Champ, Integer> {
+public interface ChampRepository extends JpaRepository<Champ, Long> {
 }
