+++ Citronix/Citronix/src/main/java/com/app/Citronix/Repository/RecolteDetailRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.Citronix.Model.Entity.DetailRecolte;
 
-public interface RecolteDetailRepository extends JpaRepository<DetailRecolte, Integer> {
+public interface RecolteDetailRepository extends JpaRepository<DetailRecolte, Long> {
 
     DetailRecolte findByRecolteIdAndArbreId(long recolteId, long arbreId);
